+++ src/features/index.ts
@@ -1,0 +1,1 @@
+export { TranslationContainer } from "./translation";
