+++ src/features/translation/ui/index.ts
@@ -1,0 +1,1 @@
+export { TranslationContainer } from "./translation-container";
