+++ src/features/translation/ui/translation-container.tsx
@@ -1,0 +1,5 @@
+import { Translation } from "./translation";
+
+export const TranslationContainer = () => {
+  return <Translation />;
+};
